requirement.txt
djangorestframework
firebase-admin
bson